make diff colors indicate equal in green and all changes in red
The default diff shows additions in green and removals in red. We want
to show the user properly recognized characters in green and the ones
that were wrong in red.

So we change color so that additions (wrongly recognized) are shown in a
darker red and missing characters (skipped recognitions) in a lighter
red.