disable start button after stop for some milliseconds
The stop button immediately restarts the exercise. Leaving the focus of the
textarea stops the exercise, makes the stop button a start button and then the
clicked signal reaches the button...

We originally tried to fix this with commit 851ce7a, but it turned out that
this does not work on mobile so we reverted it again (commit ed82b35).

This is another attempt. Since we do not have a set of events that surrounds
the mouse click that generates these events we disable the start button for
some milliseconds so the event handling can leave without restarting.

This is not the nicest solution but it seems to work.