rename lastStopAudioTime to lastAutomaticResultTime
This is used to protect the buttons from being pressed when an automatic
result was triggered by leaving the textarea. Since the purpose of those
buttons changes we do not want to click them twice when the mouse click
that leaves the area is also routed to them being clicked.

Since it is not important whether the audio was stopped but that the
automatic result was triggered, let's rename this variable.