implement anticheat for average calculation
When you have the result for one exercise a user can use that and
resubmit it until he gets a nicer average. Let's allow for each text to
be submitted only once for the average by saving hashes of the last 10
submitted texts.

This means the user needs to work through 20 exercises save their result
and then could use the first 10 to get a cheated average. That should
hopefully be too much effort. Let's see.