reveal text when using ok button without entering anything
Previously we prevented accidental full wrong texts by simply not
accepting a completely empty text and asking the user whether he wants
to enter something. Since for some trainings its just to annoying to
enter text let's just reveal the text when using the ok button with an
empty field. To make sure this can not be used for cheating we mark the
text as revealed.