slower basic letterspeed
In commit 42abac0 ("some adjustments to letterspeed and spacing",
2017-11-23) we adjusted to a higher letterspeed. It seems this was
because the author was getting more proficient hearing morse code. The
lettermix lessons should always stay for the beginner so let's adjust
back to the slower speed from before that commit.

We keep a higher inner letter speed though and adjust the spacing around
it to assemble roughly the same as 12 WPM letter and 7 WPM effective
speed.